front: drop import timetable page
