+++ front/src/applications/operationalStudies/views/Scenario/components/Timetable/Timetable.tsx
@@ -119,6 +119,7 @@ const Timetable = ({
           handleDeleteTimetableItems={handleDeleteTimetableItems}
           timetableMode={timetableMode}
           setTimetableMode={setTimetableMode}
+          upsertTimetableItems={upsertTimetableItems}
         />
         {timetableMode === 'calendar' ? (
           <Virtualizer overscan={15}>
