+++ front/src/applications/operationalStudies/views/Scenario/consts.ts
@@ -2,6 +2,5 @@ export const MANAGE_TIMETABLE_ITEM_TYPES = Object.freeze({
   none: 'NONE',
   add: 'ADD',
   edit: 'EDIT',
-  import: 'IMPORT',
   catalog: 'CATALOG',
 });
